Setup different hooks pipelines for external and internal requests
